+++ packages/react-form-composer/src/focus-on-first-field-with-error.js
@@ -1,0 +1,24 @@
+const focusOnFirstFieldWithError = (formApi) => {
+  for (const field of formApi.getFields()) {
+    const element = field.element;
+    if (field.error && element) {
+      let scrollTimeout;
+      const focusAfterScroll = e => {
+        clearTimeout(scrollTimeout);
+        scrollTimeout = setTimeout(function() {
+          if (element.focus) {
+            element.focus();
+          }
+        }, 100);
+      };    
+      focusAfterScroll();
+      window.addEventListener('scroll', focusAfterScroll);
+      if (element.scrollIntoView) {
+        element.scrollIntoView({behavior: 'smooth', block: 'end', inline: 'nearest'});
+      }
+      break;
+    }
+  }
+};
+
+export default focusOnFirstFieldWithError;